Fix #12
Remove ConnectInterface::getStatus
Remove ConnectInterface::setStatus
Remove ConnectInterface::getConnectedAt
Remove ConnectInterface::setConnectedAt
Remove ConnectInterface::getDisconnectedAt
Remove ConnectInterface::setDisconnectedAt
Remove ConnectInterface::connect
Remove ConnectInterface::disconnect
Remove ConnectInterface::STATUS_CONNECTED
Remove ConnectInterface::STATUS_DISCONNECTED

Remove ConnectionManager::connect
Remove ConnectionManager::disconnect
Rename ConnectionManager::create to connect
Rename ConnectionManager::destroy to disconnect

+++ Tests/OrmTestCase.php
@@ -82,7 +82,9 @@ class OrmTestCase extends BaseOrmTestCase
     
     public function tearDown()
     {
-        $this->schemaTool->dropSchema($this->doctrineMetadata);
+        if($this->schemaTool) {
+            $this->schemaTool->dropSchema($this->doctrineMetadata);
+        }
     }
 
     protected function getDataSet()
